+++ pdfalyzer/decorators/pdf_object_properties.py
@@ -7,7 +7,7 @@ from rich.text import Text
 
 from pdfalyzer.helpers.pdf_object_helper import pypdf_class_name
 from pdfalyzer.helpers.rich_text_helper import comma_join_txt
-from pdfalyzer.helpers.string_helper import INDENTED_JOINER, coerce_address, is_array_idx, props_string, root_address
+from pdfalyzer.helpers.string_helper import INDENTED_JOINER, coerce_address, is_array_idx, props_string_indented, root_address
 from pdfalyzer.output.styles.node_colors import get_class_style, get_class_style_dim, get_class_style_italic, get_label_style
 from pdfalyzer.util.adobe_strings import *
 from pdfalyzer.util.logging import log, log_console, log_highlighter, log_trace
@@ -162,7 +162,7 @@ class PdfObjectProperties:
             return Text(str(obj), style=get_class_style(obj))
 
     def __repr__(self) -> str:
-        return f"{type(self).__name__}(" + props_string(self, joiner=INDENTED_JOINER) + '\n)'
+        return f"{type(self).__name__}(" + props_string_indented(self) + '\n)'
 
     def __rich_without_underline__(self) -> Text:
         return self.node_label(underline=False)