+++ pdfalyzer/util/logging.py
@@ -9,12 +9,13 @@ import pypdf   # noqa: F401  # Trigger log setup?
 from rich.highlighter import ReprHighlighter
 from rich.logging import RichHandler
 from rich.theme import Theme
-# Other files import log from here to trigger log setup
-from yaralyzer.output.console import YARALYZER_THEME_DICT, console
+from yaralyzer.output.console import console
+from yaralyzer.output.theme import YARALYZER_THEME_DICT
+# Other files coult import yaralyzer's log directly but they do it from here to trigger logging setup
 from yaralyzer.util.logging import DEFAULT_LOG_HANDLER_KWARGS, log, log_console, log_trace
 
 from pdfalyzer.helpers.string_helper import regex_to_highlight_pattern, regex_to_capture_group_label
-from pdfalyzer.output.styles.node_colors import LABEL_STYLES, PARENT_STYLE, PDF_TYPE_STYLES, ClassStyle
+from pdfalyzer.output.styles.node_colors import LABEL_STYLES, PARENT_STYLE, PDF_TYPE_STYLES
 from pdfalyzer.output.styles.rich_theme import PDF_ARRAY_STYLE, PDF_DICTIONARY_STYLE
 
 LONG_ENOUGH_LABEL_STYLES = [ls for ls in LABEL_STYLES if len(ls[0].pattern) > 4]