+++ scripts/rebuild_fixtures.sh
@@ -1,3 +1,12 @@
 #!/bin/bash
 
-PYTEST_REBUILD_FIXTURES=True pytest tests/test_file_export.py
+SCRIPT_PATH=$(dirname -- "$(readlink -f -- "$0";)";)
+. "$SCRIPT_PATH/lib/project_paths.sh"
+
+
+if [ $# -eq 0 ]; then
+    print_status "\nDeleting existing fixtures rendered fixtures from $PDFALYZER_RENDERED_FIXTURES_DIR..."
+    rm "$PDFALYZER_RENDERED_FIXTURES_DIR/*.txt"
+fi
+
+PYTEST_REBUILD_FIXTURES=True pytest -vv tests/test_pdfalyze.py "$@"